Use single Formik context for frontend
We can display and access the single source of truth for packlists and items easily. Deleting a user item can now search through all the packlists and remove the item if it's used in the packlist. However this results in performance issues since Formik is rerendering all fields as the Formik "values" object changes. Additionally backend handles each Packlist as a separate document (of which each User can have multiple) and since Formik values contain all the user's packlists which could have changed so the backend needs to multiple document updates for each state update.